requirements.txt
beautifulsoup4==4.12.2
packaging==23.2
pillow==10.2.0
requests==2.31.0
urllib3==2.2.1
lxml==5.1.0
pandas==2.2.1
tqdm==4.66.2
DateTime==5.4
line-bot-sdk==3.9.0
Flask==3.0.2